Moved to DtrController.Tools the Google Holiday API John Cabugao created.
