+++ ActivityMonitoringSystem/AttendanceApp/AttendanceApp/Calendar.cs
@@ -1,19 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Threading;
-using System.IO;
-using System.Windows.Forms;
-using Google.Apis.Auth.OAuth2;
-using Google.Apis.Calendar.v3;
-using Google.Apis.Calendar.v3.Data;
-using Google.Apis.Services;
-using Google.Apis.Util.Store;
+﻿using System.Windows.Forms;
+
 
 
 namespace AttendanceApp
@@ -23,42 +9,6 @@ namespace AttendanceApp
         public Calendar()
         {
             InitializeComponent();
-            Credentials();
         }
-
-        static string[] Scopes = { CalendarService.Scope.CalendarReadonly };
-
-        public void Credentials()
-        {
-            UserCredential credential;
-
-            using (var stream =
-                new FileStream("client_secret.json", FileMode.Open, FileAccess.Read))
-            {
-                string credPath = System.Environment.GetFolderPath(
-                    System.Environment.SpecialFolder.Personal);
-                credPath = Path.Combine(credPath, ".credentials/calendar-dotnet-quickstart.json");
-
-                credential = GoogleWebAuthorizationBroker.AuthorizeAsync(
-                    GoogleClientSecrets.Load(stream).Secrets,
-                    Scopes,
-                    "user",
-                    CancellationToken.None,
-                    new FileDataStore(credPath)).Result;
-                Console.WriteLine("Credential file saved to: " + credPath);
-            }
-
-            var service = new CalendarService(new BaseClientService.Initializer()
-            {
-                HttpClientInitializer = credential,
-                ApplicationName = "Holiday Calendar",
-            });
-
-            var events = service.Events.List("en.philippines#holiday@group.v.calendar.google.com").Execute();
-
-        }
-
-
-
     }
 }
