+++ ActivityMonitoringSystem/DtrCommon/Holiday.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace DtrCommon
+{
+    public class Holiday
+    {
+        public DateTime HolidayDate { get; set; }
+        public string HolidayName { get; set; }
+    }
+}
