+++ ActivityMonitoringSystem/DtrController/packages.config
@@ -1,9 +1,15 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="EntityFramework" version="6.2.0" targetFramework="net452" />
+  <package id="Google.Apis" version="1.30.0" targetFramework="net452" />
+  <package id="Google.Apis.Auth" version="1.30.0" targetFramework="net452" />
+  <package id="Google.Apis.Calendar.v3" version="1.30.0.1020" targetFramework="net452" />
+  <package id="Google.Apis.Core" version="1.30.0" targetFramework="net452" />
+  <package id="Newtonsoft.Json" version="10.0.2" targetFramework="net452" />
   <package id="SQLite.CodeFirst" version="1.4.0.20" targetFramework="net452" />
   <package id="System.Data.SQLite" version="1.0.105.2" targetFramework="net452" />
   <package id="System.Data.SQLite.Core" version="1.0.105.2" targetFramework="net452" />
   <package id="System.Data.SQLite.EF6" version="1.0.105.2" targetFramework="net452" />
   <package id="System.Data.SQLite.Linq" version="1.0.105.2" targetFramework="net452" />
+  <package id="System.Net.Http" version="4.3.1" targetFramework="net452" />
 </packages>
